requirements.txt
asyncpg==0.27.0
dnspython==2.3.0
email-validator==2.0.0.post2
greenlet==2.0.2
idna==3.4
msgpack==1.0.5
nats-py==2.2.0
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
